+++ nodes/node_aspect_ratio_backend.py
@@ -70,5 +70,5 @@ NODE_CLASS_MAPPINGS = {
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
-    "AvataryAspectRatio": "Aspect Ratio Avatar",
+    "AvataryAspectRatio": "Aspect Ratio Avatary",
 }